Additional helper functions for local development
To reduce the overhead and chance of error when running the local server this commit automates the process.

Although I considered also modifying the HTML examples to ensure they have the correct CSS/JS reference I believe this will be unnecessary as the Dimensions content will include these in the future. We can address correcting these paths (if necessary) as new HTML content is produced from Dimensions.