Various m-list improvements
Change broadcast is only sent when changing the value or clearing the value. Component no longer emits two broadcasts as the value is cleared prior to a new value being set.

Functions more granular and refactored for more consistent naming.

If the currently selected entry is filtered out of the list, the value is now cleared.