Improvements to list selection visibility
Only scrolls to the selected value if the list is visible/when visibility changes due to interaction with a control driving the list. This improves responsiveness and fixes a 'bug' in Storybook where comboboxes and dropdowns disappear as the container scrolls in an attempt to display the selected option in the list (when the list is not currently visible).