Fix for boolean choice grid clearing
This changes the question group name, extending it to the last two parts of the name instead of the last part only. This will include, where it is present, an iteration along with a question name. If no iteration is present the current page name is captured. When rules are matched they look for the presence of the question name (supplied by the researcher) in the end of the question group, so I have not experienced any issues with testing the current smoke page.