Improvements to list visibility handling
This changes the behaviour of comboboxes and droplists to be closer to the native HTML <select> control.

Clicking now toggles the list; selecting an item now closes the list and returns focus to the control where possible; arrow keys open the list if it's closed; the enter key selects an item and closes the list.

There are still some keyboard navigation issues to conquer but this should be much closer to the desired behaviour.