Prevent placeholder messages from selecting
Placeholder messages could be selected in lists. This update stops pointer events in CSS. I considered that it may be desirable to include an additional layer of value testing via JS but concluded this may interfere with future functionality, where we submit user-supplied text without an underlying value being present.