Remove references to style-dictionary
The current version of style-dictionary references glob 11.0.1 with a known vulnerability. Although it was possible to override the include and force style-dictionary to use a more modern version the original requirement still appears in package-lock.json, flagging a warning. As we aren't currently importing tokens from an external design system I am temporarily removing the reference.